+++ src/components/Ingredient/Ingredient.jsx
@@ -13,33 +13,27 @@ import {getIngredients} from "../../services/actions/ingredients";
 
 
 const Ingredient = ({ ingredient}) => {
-  const params = useParams();
   const navigate = useNavigate();
   const location = useLocation();
- /* const [, dragRef] = useDrag({
+  //console.log(ingredient, ingredient)
+  const [, dragRef] = useDrag({
     type: ingredient.type,
     item: ingredient
   });
-*/
-  const ingredientDetailsDataSelector = (state) => state.ingredientModal.modalData;
-  const ingredientDetailsData = useSelector(ingredientDetailsDataSelector)
+
+
   const dispatch = useDispatch();
-  const handleModalClose = () => {
-    dispatch({type: DELETE_MODAL_DATA})
-    if (params.id){
-      navigate('/')
-    }
-    //setVisibleModal(false)
-  }
+
   const handleModalOpen = (ingredient) => {
-    console.log(location.state)
-    dispatch({type: SET_MODAL_DATA, payload: ingredient})
-    navigate('ingredients/'+ingredient._id, {state: {form: location.pathname}})
+    //console.log(location.state)
+    navigate('ingredients/'+ingredient._id, {state: {form: location.pathname, ingredient}})
+
+    //dispatch({type: SET_MODAL_DATA, payload: ingredient})
   }
 
   return (
     <>
-      <figure  className={style.ingredientWr}  onClick={()=>handleModalOpen(ingredient)}>
+      <figure ref={dragRef} className={style.ingredientWr}  onClick={()=>handleModalOpen(ingredient)}>
         {ingredient.count > 0 && <Counter count={ingredient.count} size="default" extraClass="m-1" />}
         <img className={`pl-4 pr-4 ${style.image}`} src={ingredient.image} alt={ingredient.name} />
         <div className={`pt-1 pb-1 ${style.priceWr}`}>
@@ -48,9 +42,7 @@ const Ingredient = ({ ingredient}) => {
         </div>
         <p className={`text text_type_main-default ${style.textCenter}`}>{ingredient.name}</p>
       </figure>
-      {ingredientDetailsData && <Modal onClose={handleModalClose} >
-        <IngredientDetails ingredient={ingredientDetailsData} />
-      </Modal>}
+
